backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
motor
pillow
aiofiles
pydantic-settings
requests
